Fix remaining narrowing conversion warnings and GCC 14 false positive
Windows MSVC fixes:
- CalChartAnimationSheet.cpp: Cast whichSheet to int for GetSheetName
- CalChartSheet.cpp: Cast enumerate index to int for GenerateCurve
- CalChartShow.cpp: Cast GetCurrentSheetNum() to int
- CalChartPrintShowToPS.cpp: Add static_cast<float> for all DPI calculations
- e7_transition_solver.cpp: Cast int division to float

Linux GCC fixes:
- CalChartContinuityToken.cpp: Suppress -Wnonnull false positive in vector<byte> operations

+++ core/CalChartAnimationSheet.cpp
@@ -265,7 +265,7 @@ auto Sheets::DebugAnimateInfoAtBeat(Beats beat) const -> std::pair<std::string, 
 {
     auto [whichSheet, newBeat] = BeatToSheetOffsetAndBeat(beat);
     std::ostringstream output;
-    output << GetSheetName(whichSheet) << " (" << static_cast<int>(whichSheet) << " of " << mSheets.size() << ")\n";
+    output << GetSheetName(static_cast<int>(whichSheet)) << " (" << static_cast<int>(whichSheet) << " of " << mSheets.size() << ")\n";
     output << "beat " << newBeat << " of " << BeatsForSheet(whichSheet) << "\n";
     auto each = mSheets.at(whichSheet).DebugAnimateInfoAtBeat(newBeat, beat == 0);
     return { output.str(), each };
